Update Node Depths
Algorithm 1(Recursion): Average case: when the tree is balanced. O(n) time | O(h) space - where n is number of nodes in the Binary Tree and h is the height of the tree.
1. Return 0 if root is null.
2. For each node, add one to the depth.
3. Call the function on left & right sub-trees and return sum of depth and results of these two function calls.

Algorithm 2(Stack and iteration): Average case: when the tree is balanced. O(n) time | O(h) space - where n is number of nodes in the Binary Tree and h is the height of the tree.
1. Create a hepler structure having the properties of a node and its depth.
2. Create a stack to hold objects of helper structure and push {root, 0} onto it.
3. While stack not empty, pop from stack:
    (a) If null, continue.
    (b) Else add the depth of current node to the result and push the node's left and right children with an incremented value of node's depth by 1.